// services/web/app/generated/blog-diagram-manifest.ts
/* Generated from public/blog/diagrams; reserve the SVG viewBox geometry. */
export const BLOG_DIAGRAM_MANIFEST = [
  {
    id: 'T01',
    locale: 'en',
    path: '/blog/diagrams/en/T01-diagram.svg',
    width: 1040,
    height: 940,
  },
  {
    id: 'T01',
    locale: 'de',
    path: '/blog/diagrams/de/T01-diagram.svg',
    width: 1040,
    height: 940,
  },
  {
    id: 'T01',
    locale: 'fr',
    path: '/blog/diagrams/fr/T01-diagram.svg',
    width: 1040,
    height: 940,
  },
  {
    id: 'T02',
    locale: 'en',
    path: '/blog/diagrams/en/T02-diagram.svg',
    width: 1040,
    height: 740,
  },
  {
    id: 'T02',
    locale: 'de',
    path: '/blog/diagrams/de/T02-diagram.svg',
    width: 1040,
    height: 740,
  },
  {
    id: 'T02',
    locale: 'fr',
    path: '/blog/diagrams/fr/T02-diagram.svg',
    width: 1040,
    height: 740,
  },
  {
    id: 'T03',
    locale: 'en',
    path: '/blog/diagrams/en/T03-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T03',
    locale: 'de',
    path: '/blog/diagrams/de/T03-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T03',
    locale: 'fr',
    path: '/blog/diagrams/fr/T03-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T04',
    locale: 'en',
    path: '/blog/diagrams/en/T04-diagram.svg',
    width: 1040,
    height: 900,
  },
  {
    id: 'T04',
    locale: 'de',
    path: '/blog/diagrams/de/T04-diagram.svg',
    width: 1040,
    height: 900,
  },
  {
    id: 'T04',
    locale: 'fr',
    path: '/blog/diagrams/fr/T04-diagram.svg',
    width: 1040,
    height: 900,
  },
  {
    id: 'T05',
    locale: 'en',
    path: '/blog/diagrams/en/T05-diagram.svg',
    width: 1040,
    height: 810,
  },
  {
    id: 'T05',
    locale: 'de',
    path: '/blog/diagrams/de/T05-diagram.svg',
    width: 1040,
    height: 810,
  },
  {
    id: 'T05',
    locale: 'fr',
    path: '/blog/diagrams/fr/T05-diagram.svg',
    width: 1040,
    height: 810,
  },
  {
    id: 'T06',
    locale: 'en',
    path: '/blog/diagrams/en/T06-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T06',
    locale: 'de',
    path: '/blog/diagrams/de/T06-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T06',
    locale: 'fr',
    path: '/blog/diagrams/fr/T06-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T07',
    locale: 'en',
    path: '/blog/diagrams/en/T07-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T07',
    locale: 'de',
    path: '/blog/diagrams/de/T07-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T07',
    locale: 'fr',
    path: '/blog/diagrams/fr/T07-diagram.svg',
    width: 1040,
    height: 890,
  },
  {
    id: 'T08',
    locale: 'en',
    path: '/blog/diagrams/en/T08-diagram.svg',
    width: 1040,
    height: 820,
  },
  {
    id: 'T08',
    locale: 'de',
    path: '/blog/diagrams/de/T08-diagram.svg',
    width: 1040,
    height: 820,
  },
  {
    id: 'T08',
    locale: 'fr',
    path: '/blog/diagrams/fr/T08-diagram.svg',
    width: 1040,
    height: 820,
  },
  {
    id: 'T09',
    locale: 'en',
    path: '/blog/diagrams/en/T09-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T09',
    locale: 'de',
    path: '/blog/diagrams/de/T09-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T09',
    locale: 'fr',
    path: '/blog/diagrams/fr/T09-diagram.svg',
    width: 1040,
    height: 850,
  },
  {
    id: 'T10',
    locale: 'en',
    path: '/blog/diagrams/en/T10-diagram.svg',
    width: 1040,
    height: 800,
  },
  {
    id: 'T10',
    locale: 'de',
    path: '/blog/diagrams/de/T10-diagram.svg',
    width: 1040,
    height: 800,
  },
  {
    id: 'T10',
    locale: 'fr',
    path: '/blog/diagrams/fr/T10-diagram.svg',
    width: 1040,
    height: 800,
  },
] as const;